Move formatter tests to formatter tests project
