+++ tests/ListPool.Formatters.Utf8Json.UnitTests/ListPoolFormatterTests.cs
@@ -1,4 +1,5 @@
-﻿using System.Linq;
+﻿using System;
+using System.Linq;
 using AutoFixture;
 using ListPool.Formatters.Utf8Json;
 using Utf8Json;
@@ -47,6 +48,18 @@ namespace ListPool.Resolvers.Utf8Json.Tests
 
             Assert.Equal(expectedItems.Count, actualItems.Count);
             Assert.All(expectedItems, expectedItem => actualItems.Contains(expectedItem));
+        }
+
+        public class CustomObject
+        {
+            public string Property { get; set; }
+        }
+
+        public sealed class CustomObjectWithListPool : CustomObject, IDisposable
+        {
+            public ListPool<int> List { get; set; }
+
+            public void Dispose() => List?.Dispose();
         }
     }
 }
