--- tests/ListPool.Resolvers.Utf8Json.Tests/CustomObject.cs
@@ -1,7 +1,0 @@
-﻿namespace ListPool.Resolvers.Utf8Json.Tests
-{
-    public class CustomObject
-    {
-        public string Property { get; set; }
-    }
-}
