+++ tests/ListPool.Resolvers.Utf8Json.Tests/ListPoolResolverTests.cs
@@ -1,4 +1,5 @@
-﻿using System.Linq;
+﻿using System;
+using System.Linq;
 using AutoFixture;
 using Utf8Json;
 using Xunit;
@@ -62,6 +63,18 @@ namespace ListPool.Resolvers.Utf8Json.Tests
             Assert.Equal(expectedItems.Count, actualObject.List.Count);
             Assert.All(expectedItems,
                 expectedItem => actualObject.List.Any(actualItem => actualItem == expectedItem));
+        }
+
+        public class CustomObject
+        {
+            public string Property { get; set; }
+        }
+
+        public sealed class CustomObjectWithListPool : CustomObject, IDisposable
+        {
+            public ListPool<int> List { get; set; }
+
+            public void Dispose() => List?.Dispose();
         }
     }
 }
